refactor: remove "index.tsx" files
feat: reorder Discord button in navbar
refactor: alphabetical ordering in CSS files
