+++ src/components/atoms/Background.tsx
