+++ src/components/organisms/Header/Underlay.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "@emotion/styled";
 import useBaseUrl from "@docusaurus/useBaseUrl";
-import Background from "../../../atoms/Background";
+import Background from "../../atoms/Background";
 
 const Mesh = styled.div<{ image?: string }>`
   position: absolute;
