+++ src/css/sablier.css
@@ -1,10 +1,15 @@
 :root {
+  --sablier-color-background: var(--sablier-color-dark000);
+  --sablier-color-backgroundLights: #484f70;
   --sablier-color-black: #000000;
+  --sablier-color-border: #2a2e41;
+  --sablier-color-borderLanding: #1e212f;
+  --sablier-color-dark: #14161f;
   --sablier-color-dark000: #14161f;
   --sablier-color-dark000-50: #14161f50;
   --sablier-color-dark050: #1a1d28;
-  --sablier-color-dark050-keep: #1a1d28;
   --sablier-color-dark050-50: #1a1d2850;
+  --sablier-color-dark050-keep: #1a1d28;
   --sablier-color-dark100: #1e212f;
   --sablier-color-dark100-50: #1e212f50;
   --sablier-color-dark150: #212433;
@@ -28,33 +33,28 @@
   --sablier-color-gray400: #8792ab;
   --sablier-color-gray500: #6a7795;
   --sablier-color-green: #00d395;
-  --sablier-color-primaryEnd: #ffb800;
-  --sablier-color-primaryMiddle: #ff9c00;
-  --sablier-color-primaryStart: #ff7300;
-  --sablier-color-red: #e52e52;
-  --sablier-color-redEnd: #c9183b;
-  --sablier-color-redDesaturated: #db3959;
-  --sablier-color-redStart: #e52e52;
-  --sablier-color-secondaryDesaturated: #266cd9;
-  --sablier-color-secondaryEnd: #00b7ff;
-  --sablier-color-secondaryMiddle: #0063ff;
-  --sablier-color-secondaryStart: #003dff;
-  --sablier-color-transparent: transparent;
-  --sablier-color-white: #ffffff;
-  --sablier-color-background: var(--sablier-color-dark000);
-  --sablier-color-backgroundLights: #484f70;
-  --sablier-color-border: #2a2e41;
-  --sablier-color-borderLanding: #1e212f;
-  --sablier-color-dark: #14161f;
   --sablier-color-inputBorder: #30354a;
   --sablier-color-light: #ffffff;
   --sablier-color-modalBackground: #1e212f;
   --sablier-color-modalDetailsBackground: #1e212f;
   --sablier-color-orange: #ff9c00;
+  --sablier-color-primaryEnd: #ffb800;
+  --sablier-color-primaryMiddle: #ff9c00;
+  --sablier-color-primaryStart: #ff7300;
   --sablier-color-purple: #805ad5;
+  --sablier-color-red: #e52e52;
+  --sablier-color-redDesaturated: #db3959;
+  --sablier-color-redEnd: #c9183b;
+  --sablier-color-redStart: #e52e52;
+  --sablier-color-secondaryDesaturated: #266cd9;
+  --sablier-color-secondaryEnd: #00b7ff;
+  --sablier-color-secondaryMiddle: #0063ff;
+  --sablier-color-secondaryStart: #003dff;
   --sablier-color-text: #e1e4ea;
   --sablier-color-thumbnailEnd: rgba(255, 156, 0, 0.01);
   --sablier-color-thumbnailStart: rgba(255, 156, 0, 0.25);
+  --sablier-color-transparent: transparent;
+  --sablier-color-white: #ffffff;
   --sablier-color-yellow: #ffb800;
 
   --sablier-size-content: 960px;
